commit - 4 - updated feed view - segregation into packages
