+++ app/src/main/java/com/example/mstcapp/navbar/feedFragment.java
@@ -1,4 +1,4 @@
-package com.example.mstcapp;
+package com.example.mstcapp.navbar;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,13 +11,19 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.mstcapp.adapters.FeedAdapter;
+import com.example.mstcapp.R;
+import com.example.mstcapp.SignInActivity;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class feedFragment extends Fragment {
     List<Integer> URLS_sample= new ArrayList<>();
+    List <String> Desc_sample=new ArrayList<>();
     Button stc_feed;
 
     RecyclerView recyclerView_feed;
@@ -31,23 +37,28 @@ public class feedFragment extends Fragment {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        URLS_sample.add(R.drawable.pp);
-        URLS_sample.add(R.drawable.pp);
-        URLS_sample.add(R.drawable.pp);
-        URLS_sample.add(R.drawable.pp);
+        URLS_sample.add(R.drawable.xbox);
+        URLS_sample.add(R.drawable.xbox);
+        URLS_sample.add(R.drawable.xbox);
+        URLS_sample.add(R.drawable.xbox);
+
+        Desc_sample.add("This the sample description for post  This is clickable and the insta post opens in new browser.");
+        Desc_sample.add("This the sample description for post  This is clickable and the insta post opens in new browser.");
+        Desc_sample.add("This the sample description for post  This is clickable and the insta post opens in new browser.");
+        Desc_sample.add("This the sample description for post  This is clickable and the insta post opens in new browser.");
 
         recyclerView_feed=(RecyclerView)getView().findViewById(R.id.recyclerview_feed);
         stc_feed=(Button)getView().findViewById(R.id.stc_home);
 
-        //layout for recycler view to have feed images
-        GridLayoutManager gridLayoutManager=new GridLayoutManager(getContext(),1);
-        recyclerView_feed.setLayoutManager(gridLayoutManager);
-        recyclerView_feed.setAdapter(new FeedAdapter(URLS_sample));
+        //layout for recycler view to have feed images or something
+
+        recyclerView_feed.setLayoutManager(new LinearLayoutManager(getContext()));
+        recyclerView_feed.setAdapter(new FeedAdapter(URLS_sample,Desc_sample));
 
         stc_feed.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivity(new Intent(getContext(),SignInActivity.class));
+                startActivity(new Intent(getContext(), SignInActivity.class));
             }
         });
