+++ app/src/main/java/com/example/mstcapp/navbar/resourcesFragment.java
@@ -1,4 +1,4 @@
-package com.example.mstcapp;
+package com.example.mstcapp.navbar;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,7 +6,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.GridLayout;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -14,6 +13,10 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.mstcapp.R;
+import com.example.mstcapp.adapters.ResourcesAdapter;
+import com.example.mstcapp.SignInActivity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -52,7 +55,7 @@ public class resourcesFragment extends Fragment {
         stc_resources.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivity(new Intent(getContext(),SignInActivity.class));
+                startActivity(new Intent(getContext(), SignInActivity.class));
             }
         });
