+++ app/src/main/java/com/example/mstcapp/resourcespage/articlelinksFragment.java
@@ -1,4 +1,4 @@
-package com.example.mstcapp;
+package com.example.mstcapp.resourcespage;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,6 +8,8 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.example.mstcapp.R;
 
 public class articlelinksFragment extends Fragment {
     public articlelinksFragment() {
