+++ app/src/main/java/com/example/mstcapp/resourcespage/resfolderFragment.java
@@ -1,4 +1,4 @@
-package com.example.mstcapp;
+package com.example.mstcapp.resourcespage;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,14 +9,16 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-public class roadmapFragment extends Fragment {
-    public roadmapFragment() {
+import com.example.mstcapp.R;
+
+public class resfolderFragment extends Fragment {
+    public resfolderFragment() {
     }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view=inflater.inflate(R.layout.fragment_roadmap,container,false);
+        View view=inflater.inflate(R.layout.fragment_resourcesfolder,container,false);
         return view;
     }
 }
