+++ app/src/main/res/layout/activity_resources_details.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".ResourcesDetails">
+    tools:context=".resourcespage.ResourcesDetails">
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
